+++ examples/with-evm-blocto-switch-chain/src/components/Button.js
@@ -1,0 +1,7 @@
+export default function BloctoButton({ children, onClick }) {
+    return (
+        <button className="btn" onClick={onClick}>
+            <span>{children}</span>
+        </button>
+    );
+}